+++ docs/release-notes.md
@@ -8,6 +8,7 @@
 - 🚀 Highlighted memory usage in status bar
 - 🆕 Add new arg to initiate dataframe for GUI rendering
 - 🆕 Add indicator for distinct values in each column for the selection box so its easier to select column for drawing
+- 🆕 Add new column on the right side panel to show distinct values for each df column
 
 ### Fixes